refactor: split _common into per-family modules
_common.py holds the specification, the command line and the driver every
maintenance-tool family runs on. Two sections of it served one family each: the
archive tools' limits, help text and three functions, and the checksum and shelf
tools' lognames, versioning and hashing.

Those two sections move out verbatim, into _archives_common.py and
_shelf_common.py beside it. Nothing changes but which module a name is read
from, and the two inner section banners, which the module headers now carry.
