+++ src/pdsfile/holdings_maintenance/_common.py
@@ -7,14 +7,15 @@
 # vocabulary -- a PDS3 volume or a PDS4 bundle -- and otherwise do the same work.
 # Everything the two halves would say twice lives here; a ToolSpec carries what
 # differs, as data.
+#
+# This module holds what every family shares: the specification, the command line,
+# and the driver. What one family shares lives beside it, in _archives_common.py
+# and _shelf_common.py.
 ##########################################################################################
 
 import argparse
-import glob
-import hashlib
 import os
 import re
-import shutil
 import sys
 from collections.abc import Callable
 from dataclasses import dataclass
@@ -182,21 +183,6 @@ def build_arg_parser(spec):
     return parser
 
 
-def reject_checksum_and_archive_paths(pdsf, path):
-    """Exit when a command-line path names checksum files or archive files.
-
-    Args:
-        pdsf: The PdsFile the path resolved to.
-        path: The absolute path the command line resolved to, for the message.
-    """
-
-    if pdsf.checksums_:
-        print('No archives for checksum files: ' + path)
-        sys.exit(1)
-
-    if pdsf.archives_:
-        print('No archives for archive files: ' + path)
-        sys.exit(1)
 
 
 def log_paths_for(pdsf, method, *args, **kwargs):
@@ -336,346 +322,3 @@ def run_main(spec, tasks, argv):
             status = 1
 
     sys.exit(status)
-
-
-##########################################################################################
-# Archive tools
-##########################################################################################
-
-# Default limits
-LOAD_DIRECTORY_INFO_LIMITS = {'info': 100}
-READ_ARCHIVE_INFO_LIMITS = {'info': 100}
-WRITE_ARCHIVE_LIMITS = {'info': -1, 'dot_': 100}
-VALIDATE_TUPLES_LIMITS = {'info': 100}
-
-ARCHIVE_DESCRIPTION = ('{progname}: Create, maintain and validate .tar.gz archives of '
-                       'PDS {unit} directory trees.')
-
-ARCHIVE_TASK_HELP = {
-    'initialize': 'Create a .tar.gz archive for a {unit}. Abort if the archive '
-                  'already exists.',
-    'reinitialize': 'Create a .tar.gz archive for a {unit}. Replace the archive if '
-                    'it already exists.',
-    'validate': 'Validate every file in a {unit} against the contents of its .tar.gz '
-                'archive. Files match if they have identical byte counts and '
-                'modification dates; file contents are not compared.',
-    'repair': 'Validate every file in a {unit} against the contents of its .tar.gz '
-              'archive. If any file has changed, write a new archive.',
-    'update': 'Search a {unit} set directory for any new {units} and create a new '
-              'archive file for each of them; do not update any pre-existing archive '
-              'files.',
-}
-
-ARCHIVE_POSITIONAL_HELP = ('The path to the root of the {unit} or {unit} set. For a '
-                           '{unit} set, all the {unit} directories inside it are '
-                           'handled in sequence.')
-
-
-def load_directory_info(spec, pdsdir, *, logger=None, limits=None):
-    """Generate a list of tuples (abspath, dirpath, nbytes, mod time)
-    recursively for the given directory tree.
-    """
-
-    if limits is None:
-        limits = {}
-
-    dirpath = pdsdir.abspath
-
-    logger = logger or pdslogger.PdsLogger.get_logger(spec.logname)
-    logger.replace_root(pdsdir.root_)
-    file_log = getattr(logger, spec.file_log_level)
-
-    merged_limits = LOAD_DIRECTORY_INFO_LIMITS.copy()
-    merged_limits.update(limits)
-    logger.open('Generating file info', dirpath, limits=merged_limits)
-
-    try:
-        lskip = spec.lskip_for(pdsdir)
-
-        tuples = [(dirpath, dirpath[lskip:], 0, 0)]
-        for (path, dirs, files) in os.walk(dirpath):
-
-            # Load files
-            for file in files:
-                abspath = os.path.join(path, file)
-
-                if file == '.DS_Store':         # skip .DS_Store files
-                    logger.ds_store('.DS_Store skipped', abspath)
-                    continue
-
-                if file.startswith('._'):       # skip dot-underscore files
-                    logger.dot_underscore('._* file skipped', abspath)
-                    continue
-
-                if BACKUP_FILENAME.match(file) or ' copy' in file:
-                    logger.error('Backup file skipped', abspath)
-                    continue
-
-                if '/.' in abspath:             # flag invisible files
-                    logger.invisible('Invisible file', abspath)
-
-                nbytes = os.path.getsize(abspath)
-                modtime = os.path.getmtime(abspath)
-                file_log('File info generated', abspath)
-
-                tuples.append((abspath, abspath[lskip:], nbytes, modtime))
-
-            # Load directories
-            for dirname in dirs:
-                abspath = os.path.join(path, dirname)
-
-                if dirname.startswith('._'):    # skip dot-underscore files
-                    logger.dot_underscore('._* directory skipped', abspath)
-                    continue
-
-                if '/.' in abspath:             # flag invisible files
-                    logger.invisible('Invisible directory', abspath)
-
-                file_log('Directory info generated', abspath)
-
-                tuples.append((abspath, abspath[lskip:], 0, 0))
-
-    except (Exception, KeyboardInterrupt) as e:
-        logger.exception(e)
-        raise
-
-    finally:
-        _ = logger.close()
-
-    return tuples
-
-
-def make_archive_filter(spec, logger, archive_invisibles):
-    """Return the tarfile member filter the archive writers add members through.
-
-    Args:
-        spec: The tool's ToolSpec.
-        logger: The logger the filter reports to.
-        archive_invisibles: True to archive invisible files, False to skip them.
-
-    Returns:
-        callable: A tarfile filter, returning the member to archive it and None to
-        skip it.
-    """
-
-    file_log = getattr(logger, spec.file_log_level)
-
-    def archive_filter(member):
-        """Internal function to filter filenames"""
-
-        # Erase user info
-        member.uid = member.gid = 0
-        member.uname = member.gname = "root"
-
-        # Check for valid file names
-        basename = os.path.basename(member.name)
-        if basename == '.DS_Store':
-            logger.ds_store('.DS_Store file skipped', member.name)
-            return None
-
-        if basename.startswith('._') or '/._' in member.name:
-            logger.dot_underscore('._* file skipped', member.name)
-            return None
-
-        if basename.startswith('.') or '/.' in member.name:
-            if archive_invisibles:
-                logger.invisible('Invisible file archived', member.name)
-                return member
-            else:
-                logger.invisible('Invisible file skipped', member.name)
-                return None
-
-        file_log('File archived', member.name)
-        return member
-
-    return archive_filter
-
-
-def validate_tuples(spec, dir_tuples, tar_tuples, *, logger=None, limits=None):
-    """Validate the directory list of tuples against the list from the tarfile.
-    """
-
-    if limits is None:
-        limits = {}
-
-    logger = logger or pdslogger.PdsLogger.get_logger(spec.logname)
-    file_log = getattr(logger, spec.file_log_level)
-
-    merged_limits = VALIDATE_TUPLES_LIMITS.copy()
-    merged_limits.update(limits)
-    logger.open('Validating file information', limits=merged_limits)
-
-    valid = True
-    try:
-        tardict = {}
-        for (abspath, dirpath, nbytes, modtime) in tar_tuples:
-            tardict[abspath] = (dirpath, nbytes, modtime)
-
-        for (abspath, dirpath, nbytes, modtime) in dir_tuples:
-            if abspath not in tardict:
-                logger.error('Missing from tar file', abspath)
-                valid = False
-
-            elif (dirpath, nbytes, modtime) != tardict[abspath]:
-
-                if nbytes != tardict[abspath][1]:
-                    logger.error('Byte count mismatch: ' +
-                                 str(nbytes) + ' (filesystem) vs. ' +
-                                 str(tardict[abspath][1]) + ' (tarfile)', abspath)
-                    valid = False
-
-                if abs(modtime - tardict[abspath][2]) > 1:
-                    logger.error('Modification time mismatch: ' +
-                                 str(modtime) + ' (filesystem) vs. ' +
-                                 str(tardict[abspath][2]) + ' (tarfile)', abspath)
-                    valid = False
-
-                del tardict[abspath]
-
-            else:
-                file_log('Validated', dirpath)
-                del tardict[abspath]
-
-        keys = list(tardict.keys())
-        keys.sort()
-        for abspath in keys:
-            logger.error('Missing from directory', abspath)
-            valid = False
-
-    except (Exception, KeyboardInterrupt) as e:
-        logger.exception(e)
-        raise
-
-    finally:
-        logger.close()
-
-    return valid
-
-
-##########################################################################################
-# Checksum and shelf file tools
-##########################################################################################
-
-# The PdsLogger name each tool kind logs under. Both flavors of a kind share one.
-CHECKSUMS_LOGNAME = 'pds.validation.checksums'
-INFOSHELF_LOGNAME = 'pds.validation.fileinfo'
-LINKSHELF_LOGNAME = 'pds.validation.links'
-
-
-@dataclass(kw_only=True)
-class VersionedFile:
-    """What move_old() needs to know about one kind of file it versions.
-
-    Attributes:
-        noun: How the file is named in the two log lines, e.g. 'Checksum file'.
-        logname: The PdsLogger name to fall back on when no logger is given.
-        companions: The extensions of the files that travel with it. Each is copied
-            beside the versioned file under the same name and its own extension. A
-            link shelf lists '.pickle', which names the shelf file itself, so that
-            file is copied twice to the one destination.
-    """
-
-    noun: str
-    logname: str
-    companions: tuple = ()
-
-
-CHECKSUM_FILE = VersionedFile(noun='Checksum file', logname=CHECKSUMS_LOGNAME)
-INFO_SHELF = VersionedFile(noun='Info shelf file', logname=INFOSHELF_LOGNAME,
-                           companions=('.py',))
-LINK_SHELF = VersionedFile(noun='Link shelf file', logname=LINKSHELF_LOGNAME,
-                           companions=('.py', '.pickle'))
-
-# The log directories a superseded checksum or shelf file is versioned into. A tool's
-# main() fills this in for each target it is about to work on; a process that never
-# calls set_log_dirs leaves it empty, and then move_old() versions nothing.
-LOGDIRS = []
-
-
-def set_log_dirs(logfiles):
-    """Record the log directories move_old() versions a superseded file into.
-
-    Args:
-        logfiles: The log file paths of the target about to be worked on. The
-            directory of each is what a superseded file is copied into.
-    """
-
-    global LOGDIRS
-    LOGDIRS = [os.path.split(logfile)[0] for logfile in logfiles]
-
-
-def next_version_dest(log_dir, prefix, ext):
-    """Return the unused <prefix>_v###<ext> path in one log directory.
-
-    ### is one past the highest version already there, and 001 when there is none.
-
-    Args:
-        log_dir: The directory the versioned copy goes in.
-        prefix: The superseded file's basename without its extension.
-        ext: That extension, including the dot.
-
-    Returns:
-        str: The path to copy to.
-    """
-
-    dest_template = log_dir + '/' + prefix + '_v???' + ext
-
-    max_version = 0
-    lskip = len(ext)
-    for version_path in glob.glob(dest_template):
-        max_version = max(max_version, int(version_path[-lskip-3:-lskip]))
-
-    return dest_template.replace('???', f'{max_version + 1:03d}')
-
-
-def move_old(path, kind, *, logger=None):
-    """Version the file a task is about to replace, into every recorded log directory.
-
-    The file is copied rather than moved, despite what the log lines say: the
-    original stays where it is and the task then overwrites it.
-
-    Args:
-        path: The file about to be replaced. Nothing happens if it does not exist,
-            or if no log directory has been recorded.
-        kind: The VersionedFile describing it.
-        logger: The logger to report through. Defaults to the kind's own.
-    """
-
-    if not os.path.exists(path):
-        return
-
-    logger = logger or pdslogger.PdsLogger.get_logger(kind.logname)
-
-    (prefix, ext) = os.path.splitext(os.path.basename(path))
-    stem = path.rpartition('.')[0]
-
-    from_logged = False
-    for log_dir in LOGDIRS:
-        dest = next_version_dest(log_dir, prefix, ext)
-        shutil.copy(path, dest)
-
-        # Both lines pass the path as the second argument, so PdsLogger renders the
-        # colon and applies the logger's root replacement to it. force=True keeps a
-        # limits cap from dropping the report of a change to the filesystem.
-        if not from_logged:
-            logger.info(kind.noun + ' moved from', path, force=True)
-            from_logged = True
-
-        logger.info(kind.noun + ' moved to', dest, force=True)
-
-        dest_stem = dest.rpartition('.')[0]
-        for companion in kind.companions:
-            shutil.copy(stem + companion, dest_stem + companion)
-
-
-# From http://stackoverflow.com/questions/3431825/-
-#       generating-an-md5-checksum-of-a-file
-
-def hashfile(fname, blocksize=65536):
-    hasher = hashlib.md5()
-
-    with open(fname, 'rb') as f:
-        for chunk in iter(lambda: f.read(blocksize), b''):
-            hasher.update(chunk)
-
-    return hasher.hexdigest()
