+++ src/pdsfile/holdings_maintenance/pds4/pds4linkshelf.py
@@ -20,9 +20,9 @@ import pdslogger
 import translator
 
 import pdsfile
-from pdsfile.holdings_maintenance import _common
+from pdsfile.holdings_maintenance import _common, _shelf_common
 
-LOGNAME = _common.LINKSHELF_LOGNAME
+LOGNAME = _shelf_common.LINKSHELF_LOGNAME
 LOGROOT_ENV = 'PDS_LOG_ROOT'
 
 # Default limits
@@ -924,7 +924,7 @@ def reinitialize(pdsdir, logger=None):
 
     # Move old file if necessary
     if os.path.exists(link_path):
-        _common.move_old(link_path, _common.LINK_SHELF, logger=logger)
+        _shelf_common.move_old(link_path, _shelf_common.LINK_SHELF, logger=logger)
 
     # Save link files
     write_linkdict(pdsdir.abspath, link_dict, logger=logger)
@@ -1004,7 +1004,7 @@ def repair(pdsdir, logger=None):
         return
 
     # Move files and write new links
-    _common.move_old(link_path, _common.LINK_SHELF, logger=logger)
+    _shelf_common.move_old(link_path, _shelf_common.LINK_SHELF, logger=logger)
     write_linkdict(pdsdir.abspath, dir_linkdict, logger=logger)
 
 def update(pdsdir,  logger=None):
@@ -1033,7 +1033,7 @@ def update(pdsdir,  logger=None):
         return
 
     # Move files and write new links
-    _common.move_old(link_path, _common.LINK_SHELF, logger=logger)
+    _shelf_common.move_old(link_path, _shelf_common.LINK_SHELF, logger=logger)
     write_linkdict(pdsdir.abspath, dir_linkdict, logger=logger)
 
 ################################################################################
@@ -1165,7 +1165,7 @@ def main():
 
             # Create all the handlers for this level in the logger
             local_handlers = []
-            _common.set_log_dirs(logfiles)
+            _shelf_common.set_log_dirs(logfiles)
             for logfile in logfiles:
                 local_handlers.append(pdslogger.file_handler(logfile))
                 logdir = os.path.split(logfile)[0]
